Fix: address feedback for Fedora package
