+++ Docker.md
@@ -31,7 +31,7 @@ docker run \
   -v /run:/run:ro \
   -v /etc/timezone:/etc/timezone:ro
   -v /etc/localtime:/etc/localtime:ro
-  -v /opt/smfc/smfc.conf:/opt/smfc/smfc.conf:ro \
+  -v /etc/smfc/smfc.conf:/etc/smfc/smfc.conf:ro \
   -e SMFC_ARGS="-l 3" \
   petersulyok/smfc
 ```
@@ -58,7 +58,7 @@ services:
     volumes:
       - /dev:/dev:ro
       - /run:/run:ro
-      - /opt/smfc/smfc.conf:/opt/smfc/smfc.conf:ro
+      - /etc/smfc/smfc.conf:/etc/smfc/smfc.conf:ro
       - /etc/timezone:/etc/timezone:ro
       - /etc/localtime:/etc/localtime:ro
     restart: unless-stopped
