+++ resources/smfc.conf
