+++ samples/smfc-sample1.conf
