+++ samples/smfc-sample2.conf
