+++ samples/smfc-sample3.conf
