+++ samples/smfc-sample4.conf
