+++ samples/smfc-sample5.conf
