+++ setup.cfg
@@ -1,0 +1,11 @@
+[metadata]
+name = smfc
+version = 3.5.0
+
+[options]
+packages =
+    smfc
+
+[options.entry_points]
+console_scripts =
+    smfc = smfc.cmd:main
