+++ smfc/cmd.py
@@ -1,0 +1,5 @@
+from .daemon import Service
+
+def main() -> None:
+    service = Service()
+    service.run()
