+++ systemd/modules-load.conf
@@ -1,0 +1,2 @@
+coretemp
+drivetemp
